server/requirements.txt
Flask>=2.2.2
flask_sqlalchemy
flask-cors
pymysql
openai==0.27.0
python-dotenv==1.0.0
torch
pypdf
scikit-learn
numpy
tiktoken
docx2txt
langchain
pydantic
typing
faiss-cpu
streamlit
streamlit_chat
boto3
flask-cors
python-docx
wkhtmltopdf
mammoth